Renaming rules to security, fixing leading whitespaces
