+++ pmd-visualforce/src/main/resources/rulesets/vf/rulesets.properties
@@ -2,4 +2,4 @@
 # BSD-style license; for more info see http://pmd.sourceforge.net/license.html
 #
 
-rulesets.filenames=rulesets/vf/basic.xml
+rulesets.filenames=rulesets/vf/security.xml
