+++ pmd-visualforce/src/test/java/net/sourceforge/pmd/lang/vf/rule/security/SecurityRulesTest.java
@@ -2,13 +2,13 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.vf.rule.basic;
+package net.sourceforge.pmd.lang.vf.rule.security;
 
 import net.sourceforge.pmd.testframework.SimpleAggregatorTst;
 
-public class BasicRulesTest extends SimpleAggregatorTst {
+public class SecurityRulesTest extends SimpleAggregatorTst {
 
-    private static final String RULESET = "vf-basic";
+    private static final String RULESET = "vf-security";
 
     @Override
     public void setUp() {
